Move pages to its own dir
