+++ docs/js/pages/game.markdown
@@ -1,0 +1,8 @@
+
+
+<!-- Start src/components/pages/game.js -->
+
+## React
+
+<!-- End src/components/pages/game.js -->
+
