+++ docs/js/pages/index.markdown
@@ -1,0 +1,8 @@
+
+
+<!-- Start src/components/pages/index.js -->
+
+## React
+
+<!-- End src/components/pages/index.js -->
+
