+++ docs/js/pages/layout.markdown
@@ -1,0 +1,8 @@
+
+
+<!-- Start src/components/pages/layout.js -->
+
+## React
+
+<!-- End src/components/pages/layout.js -->
+
