+++ docs/js/pages/settings.markdown
@@ -1,0 +1,8 @@
+
+
+<!-- Start src/components/pages/settings.js -->
+
+## React
+
+<!-- End src/components/pages/settings.js -->
+
