+++ docs/js/pages/sidebar/highscore.markdown
@@ -1,0 +1,8 @@
+
+
+<!-- Start src/components/pages/sidebar/highscore.js -->
+
+## React
+
+<!-- End src/components/pages/sidebar/highscore.js -->
+
