+++ docs/js/pages/sidebar/login.markdown
@@ -1,0 +1,8 @@
+
+
+<!-- Start src/components/pages/sidebar/login.js -->
+
+## React
+
+<!-- End src/components/pages/sidebar/login.js -->
+
