+++ src/components/pages/game.js
