+++ src/components/pages/sidebar/highscore.js
